Remove Vuex store
