+++ packages/services/f-statistics/CHANGELOG.md
@@ -3,12 +3,11 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-v1.0.0
+v0.5.0
 ------------------------------
-*Febuary 03, 2022*
+*Febuary 07, 2022*
 
 ### Added
-- Vuex module to store logs before batch publishing
 - Ability to batch publish logs based on length, interval and byte size
