--- packages/services/f-statistics/src/store/mutation-types.js
@@ -1,3 +1,0 @@
-export const ADD_LOG = 'ADD_LOG';
-export const CLEAR_LOGS = 'CLEAR_LOGS';
-
